+++ CHANGELOG.md
@@ -57,6 +57,7 @@
 ## stream-chat-android-ui-components
 ### 🐞 Fixed
 - Fixed the crash happening while editing a message with a recording attachment. [#5220](https://github.com/GetStream/stream-chat-android/pull/5220)
+- Fixed intermittent crash when opening a channel. [#5219](https://github.com/GetStream/stream-chat-android/pull/5219)
 
 ### ⬆️ Improved