+++ CHANGELOG.md
@@ -74,6 +74,10 @@
 ### ⬆️ Improved
 
 ### ✅ Added
+- Added `MessageTextFormatter` to format the message text. [#5214](https://github.com/GetStream/stream-chat-android/pull/5214)
+  * Can be overridden by `ChatTheme.messageTextFormatter`
+- Added `QuotedMessageTextFormatter` to format the quoted message text. [#5214](https://github.com/GetStream/stream-chat-android/pull/5214)
+  * Can be overridden by `ChatTheme.quotedMessageTextFormatter` 
 
 ### ⚠️ Changed